action-learning/requirements.txt
aiohttp-swagger
aiohttp[speedups]
fastdtw
networkx
numpy
opencv-python
pandas
pillow
scipy
matplotlib 
dtw-python